requirements.txt
feedparser==6.0.11
requests==2.31.0
click==8.1.7
python-dateutil==2.8.2
beautifulsoup4==4.12.2
openai==1.3.8
python-dotenv==1.0.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# New dependencies for advanced features
flask==3.0.0
anthropic==0.7.8
google-generativeai==0.3.2
schedule==1.2.0

# Production and Performance Dependencies
gunicorn==21.2.0
redis==5.0.1
psutil==5.9.6
PyJWT==2.8.0

# Development Dependencies
flask-cors==4.0.0

# Image Processing Dependencies
Pillow==10.1.0